add more comments, avoid dispatching function read flush chan when it is already in a flushing operation
